Removed the need for text processing brew info
The homebrew 'brew info' command has a JSON options that removes the need to do all the junky text processing that was previouslly done.  When package names are used this program returns quickly.  However, when keywords are specified using the --keyword option all installed brew packages must be searched.  This takes a little longer but by using the JSON option the time is much quicker than previous.