Add new vcr_hater_xref.rb
The VCR gem / technique of handling specs for API kinds of things may have been a good idea back when; maybe even now for some folks. However, within a development and test environment I think that it obscures the contract in terms of hiding outside of spec examples the expected input/output content.

I much rather see Mocks and Stubs so that my business logic is being directly tested and not the actual access of 3rd part products.

So this script just reviews the spec directory and generates two xref files for spec to cassette and cassette to spec cross-references so that I can plan the best way to extract the vcr junk.