scripts added tell; ripper-tags; updated json and rtags
ripper-tags is a nice gem that generates the ".tags" file for Ruby applications.  I updated the rtags script to make use of it.  And of course the useful_gems.rb file was updated to include ripper-tags.

I also recently did a "brew install remind" to pick up an oldie but goodie to keep track of my local reminders.  I created a new script "tell" that makes use of osascript to have the System Events application display a dialog with the reminder.